pyproject.toml
[project]
name = "heic-2-jpg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=11.3.0",
    "pillow-heif>=0.20.0",
    "tqdm>=4.66.0",
]
